# app/Http/Controllers/OperationController.php
<?php

namespace App\Http\Controllers;

use App\Repositories\OperationRepository;
use Illuminate\Http\Request;
use Symfony\Component\HttpFoundation\Response as HttpFoundationResponse;

class OperationController extends Controller
{
    public function __construct(OperationRepository $operationRepository)
    {
        $this->operationRepository = $operationRepository;
    }

    /**
    * @OA\Get(
    *     path="/api/operations/getall",
    *     tags={"operations"},
    *     summary="Get all type operations",
    *     @OA\Response(
    *         response=200,
    *         description="Successful operation",
    *         @OA\JsonContent(ref="#/components/schemas/Operation"),
    *    ),
    *     @OA\Response(
    *         response="500",
    *         description="An error has occurred."
    *     )
    * )
    */

    public function getAll(Request $request){
        return $this->getDataResponse($this->operationRepository->getAll($request), HttpFoundationResponse::HTTP_OK);
    }

    /**
    * @OA\Get(
    *     path="/api/operations/{id}",
    *     tags={"operations"},
    *     summary="Get operation by ID",
    *     @OA\Parameter(
    *         name="id",
    *         in="path",
    *         required=true,
    *         @OA\Schema(
    *             type="string"
    *         )
    *     ),
    *     @OA\Response(
    *         response=200,
    *         description="Successful operation",
    *         @OA\JsonContent(ref="#/components/schemas/Operation"),
    *    ),
    *     @OA\Response(
    *         response="404",
    *         description="Operation not found."
    *     )
    * )
    */

    public function getById(Request $request, $id){
        return $this->getDataResponse($this->operationRepository->getById($request, $id), HttpFoundationResponse::HTTP_OK);
    }

    public function create(Request $request){
        $this->validate($request, [
            'vehicle_id' => 'required|integer',
            'pending_task_id' => 'required|integer',
            'operation_type_id' => 'required|integer',
            'description' => 'required|string',
        ]);
        return $this->createDataResponse($this->operationRepository->create($request), HttpFoundationResponse::HTTP_CREATED);
    }

    /**
     * @OA\Put(
     *     path="/operations/{id}",
     *     tags={"operations"},
     *     summary="Updated operation model",
     *     @OA\RequestBody(
     *         description="Updated operation object",
     *         required=true,
     *         @OA\JsonContent(ref="#/components/schemas/Operation")
     *     ),
     *     operationId="updateOperation",
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="id that to be updated",
     *         required=true,
     *         @OA\Schema(
     *             type="string"
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Successful operation",
     *         @OA\JsonContent(ref="#/components/schemas/Operation"),
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Operation not found"
     *     ),
     * )
     */

    public function update(Request $request, $id){
        return $this->updateDataResponse($this->operationRepository->update($request, $id), HttpFoundationResponse::HTTP_OK);
    }
}