# app/Http/Controllers/ProvinceController.php
<?php

namespace App\Http\Controllers;

use Illuminate\Http\Request;
use App\Models\Province;
use App\Repositories\ProvinceRepository;
use Symfony\Component\HttpFoundation\Response as HttpFoundationResponse;

class ProvinceController extends Controller
{

    public function __construct(ProvinceRepository $provinceRepository)
    {
        $this->provinceRepository = $provinceRepository;
    }

    /**
    * @OA\Get(
    *     path="/api/provinces/getall",
    *     tags={"provinces"},
    *     summary="Get all provinces",
    *     @OA\Response(
    *         response=200,
    *         description="Successful operation",
    *         @OA\JsonContent(ref="#/components/schemas/Province"),
    *    ),
    *     @OA\Response(
    *         response="500",
    *         description="An error has occurred."
    *     )
    * )
    */

    public function getAll(Request $request){
        return $this->getDataResponse($this->provinceRepository->getAll($request), HttpFoundationResponse::HTTP_OK);
    }

    /**
    * @OA\Get(
    *     path="/api/provinces/{id}",
    *     tags={"provinces"},
    *     summary="Get province by ID",
    *     @OA\Parameter(
    *         name="id",
    *         in="path",
    *         required=true,
    *         @OA\Schema(
    *             type="string"
    *         )
    *     ),
    *     @OA\Response(
    *         response=200,
    *         description="Successful operation",
    *         @OA\JsonContent(ref="#/components/schemas/Province"),
    *    ),
    *     @OA\Response(
    *         response="404",
    *         description="Province not found."
    *     )
    * )
    */

    public function getById($id){
        return $this->getDataResponse($this->provinceRepository->getById($id), HttpFoundationResponse::HTTP_OK);
    }

    public function create(Request $request){

        $this->validate($request, [
            'region_id' => 'required|integer',
            'province_code' => 'required|string',
            'name' => 'required|string'
        ]);

        return $this->createDataResponse($this->provinceRepository->create($request), HttpFoundationResponse::HTTP_CREATED);
    }

    /**
     * @OA\Put(
     *     path="/provinces/update/{id}",
     *     tags={"provinces"},
     *     summary="Updated province",
     *     @OA\RequestBody(
     *         description="Updated province object",
     *         required=true,
     *         @OA\JsonContent(ref="#/components/schemas/Province")
     *     ),
     *     operationId="updateProvince",
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="id that to be updated",
     *         required=true,
     *         @OA\Schema(
     *             type="string"
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Successful operation",
     *         @OA\JsonContent(ref="#/components/schemas/Province"),
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Province not found"
     *     ),
     * )
     */

    public function update(Request $request, $id){
        return $this->updateDataResponse($this->provinceRepository->update($request, $id), HttpFoundationResponse::HTTP_OK);
    }

    public function delete($id){
        return $this->deleteDataResponse($this->provinceRepository->delete($id), HttpFoundationResponse::HTTP_OK);
    }
}