# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Focus",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "https://dev.api.focus.grupomobius.com"
        },
        {
            "url": "https://api.focus.grupomobius.com"
        }
    ],
    "paths": {
...
        "/incidence-images/{id}": {
            "put": {
                "tags": [
                    "incidence-images"
                ],
                "summary": "Updated incidence images",
                "operationId": "updateIncidenceImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated incidence images object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IncidenceImage"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidenceImage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Incidence image not found"
                    }
                }
            }
        },
...
        "/api/operations/getall": {
            "get": {
                "tags": [
                    "operations"
                ],
                "summary": "Get all type operations",
                "operationId": "cd4dc946bd1a5462060d0a224f6a4358",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Operation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/operations/{id}": {
            "get": {
                "tags": [
                    "operations"
                ],
                "summary": "Get operation by ID",
                "operationId": "b60abfb38a7eb5eca627f624ca13d8df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Operation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Operation not found."
                    }
                }
            }
        },
        "/operations/{id}": {
            "put": {
                "tags": [
                    "operations"
                ],
                "summary": "Updated operation model",
                "operationId": "updateOperation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated operation object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Operation"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Operation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Operation not found"
                    }
                }
            }
        },
...
        "/api/people-for-reports/getall": {
            "get": {
                "tags": [
                    "people-for-reports"
                ],
                "summary": "Get all people for reports",
                "operationId": "f60848cc9bce9ee7ba2ae71c28e4792a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PeopleForReport"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/provinces/getall": {
            "get": {
                "tags": [
                    "provinces"
                ],
                "summary": "Get all provinces",
                "operationId": "83e87a5a16564a6f09d7dc8669712ad1",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Province"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/provinces/{id}": {
            "get": {
                "tags": [
                    "provinces"
                ],
                "summary": "Get province by ID",
                "operationId": "4008b1ab9f09e5fffe2e57a1e0df6ec5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Province"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Province not found."
                    }
                }
            }
        },
        "/provinces/update/{id}": {
            "put": {
                "tags": [
                    "provinces"
                ],
                "summary": "Updated province",
                "operationId": "updateProvince",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated province object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Province"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Province"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Province not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "IncidenceImage": {
                "title": "Incidence image model",
                "description": "Incidence image model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "incidence_id": {
                        "title": "Incidence ID",
                        "description": "Incidence ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "comment_id": {
                        "title": "Comment ID",
                        "description": "Comment ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "url": {
                        "title": "URL",
                        "description": "URL",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Operation": {
                "title": "Operation model",
                "description": "Operation model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "vehicle_id": {
                        "title": "Vehicle ID",
                        "description": "Vehicle ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "pending_task_id": {
                        "title": "Pending Task ID",
                        "description": "Pending Task ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "operation_type_id": {
                        "title": "Type of Operation ID",
                        "description": "Type of Operation ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PeopleForReport": {
                "title": "People for Report model",
                "description": "People for Report model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "campa_id": {
                        "title": "Campa ID",
                        "description": "Campa ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "type_report_id": {
                        "title": "Type of Report ID",
                        "description": "Type of Report ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Province": {
                "title": "Province model",
                "description": "Province model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "region_id": {
                        "title": "Region ID",
                        "description": "Region ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "province_code": {
                        "title": "Province Code",
                        "description": "Province Code",
                        "type": "string"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}